title-pattern: match _without_ prefix, if any
In the Git for Windows project, we tried to filter out cURL's release
candidate versions via a title-pattern. However, we also specify a
prefix and that pattern was applied _after_ prefixing the title.

Not what we want. Really only match the original feed's title (or the
publication date as fallback if the feed item lacks a title).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>