Fail when issues could not be listed
As of a couple of weeks ago, GitHub seems to have serious issues with
reliability, where the issues sometimes fail to be listed.

Those issues revealed that this here Action failed to inspect the
returned HTTP status (which was previously unnecessary because it was
always 200, but this now apparently is no longer as reliable).

Let's inspect the `status` attribute to find out, and complain loudly
instead of opening duplicate issues.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>